Added the frees to the mallocs that I forgot for a long time, thanks valgrind

+++ sources/project_euler_solutions/func_exec.h
@@ -1,6 +1,12 @@
 #ifndef CROI_FUNC_ARRAY_EXEC
 #define CROI_FUNC_ARRAY_EXEC
 
-void get_function(int problem_num, bool verbose, bool testing_mode);
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <string.h>
+
+void get_function(int_fast64_t problem_num, bool verbose, bool testing_mode);
 
 #endif
